Refine subscription management flow
